Crate solana_program_runtime

Re-exports

pub use solana_rbpf;

Modules

compute_budget

compute_budget_processor

invoke_context

loaded_programs

log_collector

message_processor

prioritization_fee

stable_log

Stable program log messages

sysvar_cache

timings

Macros

declare_process_instruction

Adapter so we can unify the interfaces of built-in programs and syscalls

ic_logger_msg

Convenience macro to log a message with an Option<Rc<RefCell<LogCollector>>>

ic_msg

Convenience macro to log a message with an InvokeContext

report_execute_timings

with_mock_invoke_context