solana_program_runtime

Module compute_budget_processor

Structs

ComputeBudgetLimits

Constants

DEFAULT_INSTRUCTION_COMPUTE_UNIT_LIMIT

MAX_COMPUTE_UNIT_LIMIT

MAX_LOADED_ACCOUNTS_DATA_SIZE_BYTES

The total accounts data a transaction can load is limited to 64MiB to not break anyone in Mainnet-beta today. It can be set by set_loaded_accounts_data_size_limit instruction

Functions

process_compute_budget_instructions

Processing compute_budget could be part of tx sanitizing, failed to process these instructions will drop the transaction eventually without execution, may as well fail it early. If succeeded, the transaction’s specific limits/requests (could be default) are retrieved and returned,