solana_program_runtime::invoke_context

Struct InvokeContext

pub struct InvokeContext<'a> {
    pub transaction_context: &'a mut TransactionContext,
    pub programs_loaded_for_tx_batch: &'a LoadedProgramsForTxBatch,
    pub programs_modified_by_tx: &'a mut LoadedProgramsForTxBatch,
    pub feature_set: Arc<FeatureSet>,
    pub timings: ExecuteDetailsTimings,
    pub blockhash: Hash,
    pub lamports_per_signature: u64,
    pub syscall_context: Vec<Option<SyscallContext>>,
    /* private fields */
}

Fields
transaction_context: &'a mut TransactionContextprograms_loaded_for_tx_batch: &'a LoadedProgramsForTxBatchprograms_modified_by_tx: &'a mut LoadedProgramsForTxBatchfeature_set: Arc<FeatureSet>timings: ExecuteDetailsTimingsblockhash: Hashlamports_per_signature: u64syscall_context: Vec<Option<SyscallContext>>
Implementations

impl<'a> InvokeContext<'a>

pub fn new( transaction_context: &'a mut TransactionContext, sysvar_cache: &'a SysvarCache, log_collector: Option<Rc<RefCell<LogCollector>>>, compute_budget: ComputeBudget, programs_loaded_for_tx_batch: &'a LoadedProgramsForTxBatch, programs_modified_by_tx: &'a mut LoadedProgramsForTxBatch, feature_set: Arc<FeatureSet>, blockhash: Hash, lamports_per_signature: u64, ) -> Self

pub fn find_program_in_cache( &self, pubkey: &Pubkey, ) -> Option<Arc<LoadedProgram>>

pub fn get_environments_for_slot( &self, effective_slot: Slot, ) -> Result<&ProgramRuntimeEnvironments, InstructionError>

pub fn push(&mut self) -> Result<(), InstructionError>

Push a stack frame onto the invocation stack

pub fn pop(&mut self) -> Result<(), InstructionError>

Pop a stack frame from the invocation stack

pub fn get_stack_height(&self) -> usize

Current height of the invocation stack, top level instructions are height solana_sdk::instruction::TRANSACTION_LEVEL_STACK_HEIGHT

pub fn native_invoke( &mut self, instruction: StableInstruction, signers: &[Pubkey], ) -> Result<(), InstructionError>

Entrypoint for a cross-program invocation from a builtin program

pub fn prepare_instruction( &mut self, instruction: &StableInstruction, signers: &[Pubkey], ) -> Result<(Vec<InstructionAccount>, Vec<IndexOfAccount>), InstructionError>

Helper to prepare for process_instruction()

pub fn process_instruction( &mut self, instruction_data: &[u8], instruction_accounts: &[InstructionAccount], program_indices: &[IndexOfAccount], compute_units_consumed: &mut u64, timings: &mut ExecuteTimings, ) -> Result<(), InstructionError>

Processes an instruction and returns how many compute units were used

pub fn get_log_collector(&self) -> Option<Rc<RefCell<LogCollector>>>

Get this invocation’s LogCollector

pub fn consume_checked(&self, amount: u64) -> Result<(), Box<dyn Error>>

Consume compute units

pub fn mock_set_remaining(&self, remaining: u64)

Set compute units

Only use for tests and benchmarks

pub fn get_compute_budget(&self) -> &ComputeBudget

Get this invocation’s compute budget

pub fn get_sysvar_cache(&self) -> &SysvarCache

Get cached sysvars

pub fn get_check_aligned(&self) -> bool

pub fn set_syscall_context( &mut self, syscall_context: SyscallContext, ) -> Result<(), InstructionError>

pub fn get_syscall_context(&self) -> Result<&SyscallContext, InstructionError>

pub fn get_syscall_context_mut( &mut self, ) -> Result<&mut SyscallContext, InstructionError>

pub fn get_traces(&self) -> &Vec<Vec<[u64; 12]>>

Return a references to traces