Module solana_program_runtime::sysvar_cache::get_sysvar_with_account_check

These methods facilitate a transition from fetching sysvars from keyed accounts to fetching from the sysvar cache without breaking consensus. In order to keep consistent behavior, they continue to enforce the same checks as solana_sdk::keyed_account::from_keyed_account despite dynamically loading them instead of deserializing from account data.

Functions

check_sysvar_account 🔒

clock

recent_blockhashes

rent

slot_hashes

stake_history