Crate solana_program_runtime

Modules

accounts_data_meter

The accounts data space has a maximum size it is permitted to grow to. This module contains the constants and types for tracking and metering the accounts data space during program runtime.

compute_budget

executor_cache

invoke_context

loaded_programs

log_collector

pre_account

prioritization_fee

stable_log

Stable program log messages

sysvar_cache

timings

Macros

declare_process_instruction

Adapter so we can unify the interfaces of built-in programs and syscalls

ic_logger_msg

Convenience macro to log a message with an Option<Rc<RefCell<LogCollector>>>

ic_msg

Convenience macro to log a message with an InvokeContext

report_execute_timings

with_mock_invoke_context